+++ instagramsdk/src/main/java/me/aravi/instapi/bean/post/MusicMetadata.java
@@ -54,7 +54,7 @@ public class MusicMetadata {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        sb.append(MusicMetadata.class.getName()).append('@').append(Integer.toHexString(System.identityHashCode(this))).append('[');
+        sb.append(MusicMetadata.class.getName()).append('@').append(Long.toHexString(System.identityHashCode(this))).append('[');
         sb.append("musicCanonicalId");
         sb.append('=');
         sb.append(((this.musicCanonicalId == null)?"<null>":this.musicCanonicalId));
